refactor: refactored opencl algorithms in order to support non-gpu able systems
