+++ examples/CMakeLists.txt
@@ -13,4 +13,4 @@ add_demo(example_load_network)
 add_demo(example_network)
 add_demo(example_named_network)
 add_demo(example_interact_network)
-add_demo(example_gpu_inference)
+add_demo(example_cuda_inference)
