+++ test/CMakeLists.txt
@@ -5,6 +5,7 @@ FetchContent_Declare(
         googletest
         GIT_REPOSITORY https://github.com/google/googletest.git
         GIT_TAG        e2239ee6043f73722e7aa812a459f54a28552929 # release-1.11.0
+
 )
 
 if (WIN32)
