+++ test/evidence_test.cpp
@@ -6,14 +6,18 @@
 #include <baylib/network/bayesian_net.hpp>
 #include <baylib/smile_utils/smile_utils.hpp>
 #include <baylib/inference/gibbs_sampling.hpp>
-#include <baylib/inference/logic_sampling.hpp>
 #include <baylib/inference/likelihood_weighting.hpp>
 #include <baylib/inference/rejection_sampling.hpp>
-#include <baylib/inference/adaptive_importance_sampling.hpp>
 #include <baylib/network/bayesian_utils.hpp>
-#ifdef CUDA_CMP_FOUND
-#include <baylib/inference/logic_sampling_cuda.hpp>
-#include <baylib/inference/likelihood_weighting_cuda.hpp>
+
+#ifdef BAYLIB_CUDA
+#include "baylib/inference/cuda/logic_sampling_cuda.hpp"
+#include "baylib/inference/cuda/likelihood_weighting_cuda.hpp"
+#endif
+
+#ifdef BAYLIB_OPENCL
+#include "baylib/inference/opencl/logic_sampling_opencl.hpp"
+#include <baylib/inference/opencl/adaptive_importance_sampling_opencl.hpp>
 #endif
 
 #define THREADS std::thread::hardware_concurrency()
@@ -32,14 +36,16 @@ using bnet = baylib::bayesian_net<Variable>;
 template<typename Probability, class Variable>
         std::vector<baylib::marginal_distribution<Probability>> get_results(const bnet<Variable> &bn){
             std::vector<baylib::marginal_distribution<Probability>> results{
-                logic_sampling<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
-                gibbs_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                likelihood_weighting<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                rejection_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                adaptive_importance_sampling<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
-#ifdef CUDA_CMP_FOUND
-                logic_sampling_cuda<bnet<Variable>>(bn, SAMPLES).make_inference(),
-                likelihood_weighting_cuda<bnet<Variable>>(bn, SAMPLES).make_inference()
+                    gibbs_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+                    likelihood_weighting<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+                    rejection_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+#ifdef BAYLIB_CUDA
+                    logic_sampling_cuda<bnet<Variable>>(bn, SAMPLES).make_inference(),
+                    likelihood_weighting_cuda<bnet<Variable>>(bn, SAMPLES).make_inference(),
+#endif
+#ifdef BAYLIB_OPENCL
+                    logic_sampling_opencl<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
+                    adaptive_importance_sampling_opencl<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
 #endif
             };
             return results;
@@ -48,12 +54,14 @@ template<typename Probability, class Variable>
 template<typename Probability, class Variable>
         std::vector<baylib::marginal_distribution<Probability>> get_results_heavy(const bnet<Variable> &bn){
             std::vector<baylib::marginal_distribution<Probability>> results{
-                likelihood_weighting<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                gibbs_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                rejection_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
-                adaptive_importance_sampling<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
-#ifdef CUDA_CMP_FOUND
-                likelihood_weighting_cuda<bnet<Variable>>(bn, SAMPLES).make_inference()
+                    likelihood_weighting<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+                    gibbs_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+                    rejection_sampling<bnet<Variable>>(bn, SAMPLES, THREADS).make_inference(),
+#ifdef BAYLIB_OPENCL
+                    adaptive_importance_sampling_opencl<bnet<Variable>>(bn, SAMPLES, MEMORY).make_inference(),
+#endif
+#ifdef BAYLIB_CUDA
+                    likelihood_weighting_cuda<bnet<Variable>>(bn, SAMPLES).make_inference()
 #endif
             };
             return results;
